+++ interface/frontend/src/app/app.module.ts
@@ -5,12 +5,15 @@ import { RouterModule, RouterOutlet } from "@angular/router";
 import { HttpClientModule } from "@angular/common/http";
 import { NgOptimizedImage } from "@angular/common";
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
+import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 
 import { MatToolbarModule } from "@angular/material/toolbar";
 import { MatButtonModule } from "@angular/material/button";
 import { MatIconModule } from "@angular/material/icon";
 import { MatSidenavModule } from "@angular/material/sidenav";
 import { MatListModule } from "@angular/material/list";
+
+import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
 
 import { NgxSimpleTextEditorModule } from 'ngx-simple-text-editor';
 import { QuillModule } from "ngx-quill";
@@ -19,7 +22,7 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { CustomSidenavComponent } from './components/custom-sidenav/custom-sidenav.component';
 import { HomeComponent } from './components/pages/home/home.component';
-import { CalendarComponent } from './components/pages/calendar/calendar/calendar.component';
+import { CalendarComponent } from './components/pages/calendar/calendar.component';
 
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
@@ -43,9 +46,11 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
     MatListModule,
     NgOptimizedImage,
     RouterModule,
+    BrowserAnimationsModule,
     NgxSimpleTextEditorModule,
     QuillModule.forRoot(),
-    NgbModule
+    NgbModule,
+    FontAwesomeModule
   ],
   providers: [
     provideClientHydration(),