+++ interface/frontend/src/app/services/calendar/day/day.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { DayServiceService } from './day-service.service';
+import { DayServiceService } from './day.service';
 
 describe('DayServiceService', () => {
   let service: DayServiceService;