+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/places/Capital.java
@@ -1,19 +1,32 @@
 package com.testament.veltahleon.model.entities.places;
 
 import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-@EqualsAndHashCode(callSuper = true)
 @Entity
-@Data
+@Getter
+@Setter
 @AllArgsConstructor
 @NoArgsConstructor
 @Table(name = "capitals")
+@EqualsAndHashCode
 @ToString
-public class Capital extends Location {
+public class Capital {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @NotNull(message = "Name cannot be null!")
+    @NotBlank(message = "Name cannot be blank!")
+    @NotEmpty(message = "Name cannot be empty!")
+    @Column(name = "name")
+    private String name;
+
+    @Column(name = "description", columnDefinition = "text")
+    private String description;
 }