+++ veltahleon/src/main/java/com/testament/veltahleon/services/ifc/divination/ConstellationService.java
@@ -12,5 +12,4 @@ public interface ConstellationService {
     Boolean deleteConstellationByID(Long id);
     Constellation saveConstellation(Constellation constellation);
     Constellation updateConstellation(Long id, Constellation constellation);
-    String addImageURL(String fileName);
 }