+++ veltahleon/src/main/java/com/testament/veltahleon/services/ifc/places/CapitalService.java
@@ -10,7 +10,7 @@ public interface CapitalService {
     Collection<Capital> getCapitals();
     Capital getCapitalByID(Long id);
     Capital getCapitalByName(String name);
-    Capital getCapitalByNationName(String name);
+    //Capital getCapitalByNationName(String name);
     Boolean deleteCapitalByID(Long id);
     Capital saveCapital(Capital capital);
     Capital updateCapital(Long id, Capital capital);